pyproject.toml
[project]
name = "vlm"
version = "0.1.0"
description = "Parameter-efficient adaptation of compact vision-language models"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.47.0",
    "trl>=0.13.0",
    "peft>=0.14.0",
    "bitsandbytes>=0.45.0",
    "accelerate>=1.2.0",
    "datasets>=3.2.0",
    "Pillow>=11.0.0",
    "tensorboard>=2.18.0",
    "evaluate>=0.4.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "ipython>=8.30.0",
]

[tool.uv]
# Use the PyTorch CUDA 12.4 index only for torch/torchvision.
# The PyTorch index ships stale copies of unrelated packages (requests, markupsafe)
# that break resolution, so we scope it via tool.uv.sources.
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
